gas_tracker

+++ app/src/main/java/com/wtb/dashTracker/util/CSVUtils.kt
@@ -20,21 +20,22 @@ import android.content.Context
 import android.content.Intent
 import android.content.Intent.EXTRA_STREAM
 import android.widget.Toast
-import androidx.core.content.ContextCompat.startActivity
 import androidx.core.content.FileProvider
-import com.github.doyaaaaaken.kotlincsv.dsl.csvReader
 import com.github.doyaaaaaken.kotlincsv.dsl.csvWriter
 import com.wtb.dashTracker.MainActivity
 import com.wtb.dashTracker.R
-import com.wtb.dashTracker.database.models.DashEntry
+import com.wtb.dashTracker.database.models.*
 import com.wtb.dashTracker.database.models.DashEntry.Companion.asList
-import com.wtb.dashTracker.database.models.DataModel
-import com.wtb.dashTracker.database.models.Weekly
+import com.wtb.dashTracker.database.models.Expense.Companion.asList
+import com.wtb.dashTracker.database.models.ExpensePurpose.Companion.asList
 import com.wtb.dashTracker.database.models.Weekly.Companion.asList
-import com.wtb.dashTracker.repository.Repository
-import com.wtb.dashTracker.ui.dialog_confirm_delete.ConfirmationDialog
+import com.wtb.dashTracker.ui.dialog_confirm.ConfirmationDialogExport
+import com.wtb.dashTracker.ui.dialog_confirm.ConfirmationDialogImport
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import java.io.*
+import java.io.BufferedInputStream
+import java.io.BufferedOutputStream
+import java.io.File
+import java.io.FileInputStream
 import java.time.LocalDate
 import java.util.zip.ZipEntry
 import java.util.zip.ZipOutputStream
@@ -42,10 +43,14 @@ import java.util.zip.ZipOutputStream
 
 @ExperimentalCoroutinesApi
 class CSVUtils {
-    private val repository: Repository
-        get() = Repository.get()
 
-    fun export(entries: List<DashEntry>?, weeklies: List<Weekly>?, ctx: Context) {
+    fun export(
+        entries: List<DashEntry>?,
+        weeklies: List<Weekly>?,
+        expenses: List<Expense>,
+        purposes: List<ExpensePurpose>,
+        ctx: Context
+    ) {
         fun generateFile(context: Context, fileName: String): File? {
             val csvFile = File(context.filesDir, fileName)
             csvFile.createNewFile()
@@ -71,6 +76,24 @@ class CSVUtils {
                 writeRow(Weekly.headerList)
                 weeklies?.forEach { weekly ->
                     writeRow(weekly.asList())
+                }
+            }
+        }
+
+        fun exportExpenses(csvFile: File, expenses: List<Expense>?) {
+            csvWriter().open(csvFile, append = false) {
+                writeRow(Expense.headerList)
+                expenses?.forEach { expense ->
+                    writeRow(expense.asList())
+                }
+            }
+        }
+
+        fun exportPurposes(csvFile: File, purposes: List<ExpensePurpose>?) {
+            csvWriter().open(csvFile, append = false) {
+                writeRow(ExpensePurpose.headerList)
+                purposes?.forEach { purpose ->
+                    writeRow(purpose.asList())
                 }
             }
         }
@@ -107,30 +130,25 @@ class CSVUtils {
 
         val dailyFile: File? = generateFile(ctx, getEntriesFileName())
         val weeklyFile: File? = generateFile(ctx, getWeekliesFileName())
+        val expenseFile: File? = generateFile(ctx, getExpensesFileName())
+        val purposeFile: File? = generateFile(ctx, getPurposesFileName())
 
-        if (dailyFile != null && weeklyFile != null) {
+        if (dailyFile != null && weeklyFile != null && expenseFile != null && purposeFile != null) {
             exportEntries(dailyFile, entries)
             exportWeeklies(weeklyFile, weeklies)
+            exportExpenses(expenseFile, expenses)
+            exportPurposes(purposeFile, purposes)
 
-            val zipFile: File = zipFiles(ctx, dailyFile, weeklyFile)
+            val zipFile: File = zipFiles(ctx, dailyFile, weeklyFile, expenseFile, purposeFile)
             dailyFile.delete()
             weeklyFile.delete()
+            expenseFile.delete()
+            purposeFile.delete()
 
             (ctx as MainActivity).runOnUiThread {
-                ConfirmationDialog(
-                    text = R.string.confirm_export,
-                    requestKey = "confirmExport",
-                    message = "Confirm Export",
-                    posButton = R.string.label_action_export_csv,
-                    posAction = {
-                        startActivity(
-                            ctx,
-                            Intent.createChooser(getSendFilesIntent(zipFile), null),
-                            null
-                        )
-                    },
-                    negButton = R.string.cancel,
-                    negAction = { },
+                ConfirmationDialogExport(
+                    ctx,
+                    getSendFilesIntent(zipFile)
                 ).show(ctx.supportFragmentManager, null)
             }
         } else {
@@ -146,32 +164,16 @@ class CSVUtils {
 
     fun import(ctx: Context) {
         (ctx as MainActivity).runOnUiThread {
-            ConfirmationDialog(
-                text = R.string.confirm_import,
-                requestKey = "confirmImportEntry",
-                message = "Confirm Import",
-                posButton = R.string.label_action_import_csv,
-                posAction = {
-                    ctx.contentZipLauncher.launch("application/zip")
-                },
-                negButton = R.string.cancel,
-                negAction = { },
-            ).show(ctx.supportFragmentManager, null)
+            ConfirmationDialogImport(ctx).show(ctx.supportFragmentManager, null)
         }
     }
-
-    private fun <T : DataModel> getModelsFromCsv(
-        path: InputStream?,
-        function: (Map<String, String>) -> T
-    ): List<T>? =
-        path?.let { inStream ->
-            csvReader().readAllWithHeader(inStream).map { function(it) }
-        }
 
     companion object {
         const val FILE_ZIP = "dash_tracker_"
         const val FILE_ENTRIES = "dash_tracker_entries_"
         const val FILE_WEEKLIES = "dash_tracker_weeklies_"
+        const val FILE_EXPENSES = "dash_tracker_expenses_"
+        const val FILE_PURPOSES = "dash_tracker_purposes_"
 
         private fun getEntriesFileName() =
             "$FILE_ENTRIES${LocalDate.now().toString().replace('-', '_')}.csv"
@@ -181,6 +183,12 @@ class CSVUtils {
 
         private fun getWeekliesFileName() =
             "$FILE_WEEKLIES${LocalDate.now().toString().replace('-', '_')}.csv"
+
+        private fun getExpensesFileName() =
+            "$FILE_EXPENSES${LocalDate.now().toString().replace('-', '_')}.csv"
+
+        private fun getPurposesFileName() =
+            "$FILE_PURPOSES${LocalDate.now().toString().replace('-', '_')}.csv"
     }
 }
